pyproject.toml
[tool.poetry]
name = "aws-project"
version = "0.1.0"
description = ""
authors = ["Lucas <lucasdastre10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
boto3 = "^1.34.162"
pyarrow = "^17.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
